[#5249] Delay server now creates its own PID file.
